Final commit... for now
At this time, it looks like Piltover Archive has the deck API open, but the card API closed.

And Riftmana has seems to have the API hidden away from the public when doing some quick scanning.

As a result, this is the best that can be done for now...

The changes this time were adding an argument in the CLI to choose the image server for the desklists and removing the initial Pixelborn and Piltover Archive decklist options from the code, since neither of those can be approached until a card API is found to be usable.